Explained predictor variance

Extraction and/or computation of explained variances for various object classes in the multiblock package.

Usage

explvar(object)

Arguments

object

An object fitted using a method from the multiblock package

Value

A vector of component-wise explained variances for predictors.

Examples

data(potato)
so <- sopls(Sensory ~ Chemical + Compression, data=potato, ncomp=c(10,10), 
            max_comps=10)
explvar(so)
#>        0,0        0,1        0,2        0,3        0,4        0,5        0,6 
#>  0.0000000 45.8687627  9.6453264  7.3898305  3.5545689  1.2708076  7.2218549 
#>        0,7        0,8        0,9       0,10        1,0        1,1        1,2 
#>  3.3122521  1.5830365  2.5404624  0.7899845 34.2017249 29.9203640  7.5862050 
#>        1,3        1,4        1,5        1,6        1,7        1,8        1,9 
#>  4.5354153  1.9562759  1.2891286  3.8823352  1.0455730  1.1736034  0.4786749 
#>        2,0        2,1        2,2        2,3        2,4        2,5        2,6 
#>  8.5567408 29.8494182  6.7924606  2.0046763  1.2816857  2.2955734  3.1794417 
#>        2,7        2,8        3,0        3,1        3,2        3,3        3,4 
#>  0.7035895  0.9788461 20.9999933 14.8311902  5.9640780  2.2114124  0.9630738 
#>        3,5        3,6        3,7        4,0        4,1        4,2        4,3 
#>  2.3612130  1.7170928  0.6543032  6.3303082 10.2624882  6.7734592  1.7625131 
#>        4,4        4,5        4,6        5,0        5,1        5,2        5,3 
#>  0.8944426  2.2524447  1.3335044  5.1247451  9.0556886  6.8925550  1.3579712 
#>        5,4        5,5        6,0        6,1        6,2        6,3        6,4 
#>  0.9480858  1.9343360  2.0713113  8.9383967  6.9547883  1.3697888  1.0793572 
#>        7,0        7,1        7,2        7,3        8,0        8,1        8,2 
#>  2.2840233  8.8631448  7.0319143  1.1838720  3.2504266  5.3411760  7.8007596 
#>        9,0        9,1       10,0 
#>  2.6747591  4.7191749  1.5823723