multiblock

A collection of methods for analysis of data sets with more than two blocks of data.

Unsupervised methods:

SCA - Simultaneous Component Analysis (sca)

GCA - Generalized Canonical Analysis (gca)

GPA - Generalized Procrustes Analysis (gpa)

MFA - Multiple Factor Analysis (mfa)

PCA-GCA (pcagca)

DISCO - Distinctive and Common Components with SCA (disco)

HPCA - Hierarchical Principal component analysis (hpca)

MCOA - Multiple Co-Inertia Analysis (mcoa)

JIVE - Joint and Individual Variation Explained (jive)

STATIS - Structuration des Tableaux à Trois Indices de la Statistique (statis)

HOGSVD - Higher Order Generalized SVD (hogsvd)

Design based methods:

ASCA - Anova Simultaneous Component Analysis (asca)

Supervised methods:

MB-PLS - Multiblock Partial Least Squares (mbpls)

sMB-PLS - Sparse Multiblock Partial Least Squares (smbpls)

SO-PLS - Sequential and Orthogonalized PLS (sopls)

PO-PLS - Parallel and Orthogonalized PLS (popls)

ROSA - Response Oriented Sequential Alternation (rosa)

mbRDA - Multiblock Redundancy Analysis (mbrda)

Complex methods:

L-PLS - Partial Least Squares in L configuration (lpls)

SO-PLS-PM - Sequential and Orthogonalised PLS Path Modelling (sopls_pm)

Single- and two-block methods:

PCA - Principal Component Analysis (pca)

PCR - Principal Component Regression (pcr)

PLSR - Partial Least Squares Regression (plsr)

CCA - Canonical Correlation Analysis (cca)

IFA - Interbattery Factor Analysis (ifa)

GSVD - Generalized SVD (gsvd)

Datasets:

Sensory assessment of candies (candies)

Sensory, rheological, chemical and spectroscopic analysis of potatoes (potato)

Cheese production process data (process)

Data simulated to have certain characteristics (simulated)

Wines of Val de Loire (wine)

Utility functions:

Block-wise indexable data.frame (block.data.frame)

Dummy-code a vector (dummycode)

See also