Result functions for SO-PLS models

Standard result functions for SO-PLS (sopls).

# S3 method for sopls
predict(
  object,
  newdata,
  ncomp = object$ncomp,
  comps = object$ncomp,
  type = c("response", "scores"),
  na.action = na.pass,
  ...
)

# S3 method for sopls
coef(
  object,
  ncomp = object$ncomp,
  comps = object$ncomp,
  intercept = FALSE,
  ...
)

# S3 method for sopls
print(x, ...)

# S3 method for sopls
summary(
  object,
  what = c("all", "validation", "training"),
  digits = 4,
  print.gap = 2,
  ...
)

classify(object, ...)

# S3 method for sopls
classify(object, classes, newdata, ncomp, LQ = "LDA", ...)

# S3 method for sopls
loadings(object, ncomp = "all", block = 1, ...)

# S3 method for sopls
loadingplot(object, ncomp = "all", comps = c(1, 2), block = 1, ...)

# S3 method for sopls
scores(object, ncomp = "all", block = 1, ...)

# S3 method for sopls
scoreplot(object, ncomp = "all", comps = c(1, 2), block = 1, ...)

# S3 method for sopls
R2(object, estimate, newdata, ncomp = "all", individual = FALSE, ...)

# S3 method for sopls
RMSEP(object, estimate, newdata, ncomp = "all", individual = FALSE, ...)

Arguments

object
A sopls object.

newdata
Optional new data with the same types of predictor blocks as the ones used for fitting the object.

ncomp
An integer giving the number of components to use apply.

comps
An integer vector giving the exact components to apply.

type
Character indicating which type of explained variance to compute (default = "train", alternative = "CV").

na.action
Function determining what to do with missing values in newdata.

...
Additional arguments. Currently not implemented.

intercept
A logical indicating if coefficients for the intercept should be included (default = FALSE).

x
A sopls object.

what
A character indicating if summary should include all, validation or training.

digits
The number of digits used for printing.

print.gap
Gap between columns when printing.

classes
A character vector of class labels.

LQ
A character indicating if 'max' (maximum score value), 'lda' or 'qda' should be used when classifying.

block
An integer indicating which block to use.

estimate
A character indicating if 'train', 'CV' or 'test' results should be displayed.

individual
A logical indicating if results for individual responses should be displayed.

Value

Returns depend on method used, e.g. predict.sopls returns predicted responses or scores depending on inputs, coef.sopls return regression coefficients.

References

Jørgensen K, Mevik BH, Næs T. Combining designed experiments with several blocks of spectroscopic data. Chemometr Intell Lab Syst. 2007;88(2): 154–166.

See also

Examples

data(potato)
mod <- sopls(Sensory[,1] ~ ., data = potato[c(1:3,9)], ncomp = 5, subset = 1:20)
testset <- potato[-(1:20),]; testset$Sensory <- testset$Sensory[,1,drop=FALSE]
predict(mod, testset, comps=c(2,1,2))

#> , , 1,0,0
#> 
#>      Sensory[, 1]
#> [1,]     3.922829
#> [2,]     3.801854
#> [3,]     6.409450
#> [4,]     2.960830
#> [5,]     4.433857
#> [6,]     3.223006
#> 
#> , , 2,0,0
#> 
#>      Sensory[, 1]
#> [1,]     4.062985
#> [2,]     4.238679
#> [3,]     7.178207
#> [4,]     2.477011
#> [5,]     3.319856
#> [6,]     2.904183
#> 
#> , , 2,1,0
#> 
#>      Sensory[, 1]
#> [1,]     4.169790
#> [2,]     4.471989
#> [3,]     7.282654
#> [4,]     2.426714
#> [5,]     3.273465
#> [6,]     2.898753
#> 
#> , , 2,1,1
#> 
#>      Sensory[, 1]
#> [1,]    5.2214087
#> [2,]    5.1427034
#> [3,]    9.7861678
#> [4,]   -0.7742392
#> [5,]    0.9645707
#> [6,]    2.2666501
#> 
#> , , 2,1,2
#> 
#>      Sensory[, 1]
#> [1,]    5.4727706
#> [2,]    5.5669734
#> [3,]   11.3237645
#> [4,]   -2.3105720
#> [5,]    0.2755925
#> [6,]    2.1649361
#> 

#dim(coef(mod, ncomp=5)) # <variables x responses x components>
print(mod)

#> Sequential and Orthogonalized Partial Least Squares, fitted with the PKPLS algorithm.
#> Call:
#> sopls(formula = Sensory[, 1] ~ ., ncomp = 5, data = potato[c(1:3,     9)], subset = 1:20)

summary(mod)

#> Data: 	X dimension: 20 0 
#> 	Y dimension: 20 1
#> Fit method: PKPLS
#> Number of components considered: 5
#> TRAINING: % variance explained
#>               0,0,0  0,0,1  0,0,2  0,0,3  0,0,4  0,0,5  0,1,0  0,1,1  0,1,2
#> X                 0  20.86  28.31  36.64  44.50  49.32  43.07  48.93  57.63
#> Sensory[, 1]      0  67.18  76.02  79.52  82.14  86.78  31.94  77.34  78.82
#>               0,1,3  0,1,4  0,2,0  0,2,1  0,2,2  0,2,3  0,3,0  0,3,1  0,3,2
#> X             64.80  72.01  53.34   58.2  63.49  70.46  64.51  67.17  68.88
#> Sensory[, 1]  81.95  85.33  57.59   79.4  84.80  86.19  64.12  85.01  86.62
#>               0,4,0  0,4,1  0,5,0  1,0,0  1,0,1  1,0,2  1,0,3  1,0,4  1,1,0
#> X             67.33  68.65  71.36  26.79  31.20  43.68  46.20  52.01  61.24
#> Sensory[, 1]  74.00  83.04  80.83  72.84  91.32  92.82  95.22  95.51  76.08
#>               1,1,1  1,1,2  1,1,3  1,2,0  1,2,1  1,2,2  1,3,0  1,3,1  1,4,0
#> X             66.52  71.52  74.07  67.65  73.86  76.42  77.02  78.22  78.39
#> Sensory[, 1]  88.73  93.14  95.49  79.16  88.85  93.93  81.81  93.93  85.47
#>               2,0,0  2,0,1  2,0,2  2,0,3  2,1,0  2,1,1  2,1,2  2,2,0  2,2,1
#> X             34.87  37.53  48.20  50.75  68.53  70.72  76.16  75.89  78.58
#> Sensory[, 1]  86.39  94.20  94.31  95.79  87.33  93.13  94.50  88.70  93.06
#>               2,3,0  3,0,0  3,0,1  3,0,2  3,1,0  3,1,1  3,2,0  4,0,0  4,0,1
#> X             79.17  59.03  60.65  67.95  69.06  71.96  78.92  70.41  72.12
#> Sensory[, 1]  92.41  87.55  94.19  94.84  88.70  93.08  90.25  88.65  95.40
#>               4,1,0  5,0,0
#> X             77.16  77.24
#> Sensory[, 1]  90.66  90.58

#blockexpl(mod)
#print(blockexpl(mod), compwise=TRUE)