Response Oriented Sequential Alternation - ROSA

ROSA is an opportunistic method sequentially selecting components from whichever block explains the response most effectively. It can be formulated as a PLS model on concatenated input block with block selection per component. This implementation adds several options that are not described in the literature. Most importantly, it opens for internal validation in the block selection process, making this more robust. In addition it handles design blocks explicitly, enables classification and secondary responses (CPLS), and definition of common components.

rosa(
  formula,
  ncomp,
  Y.add,
  common.comp = 1,
  data,
  subset,
  na.action,
  scale = FALSE,
  weights = NULL,
  validation = c("none", "CV", "LOO"),
  internal.validation = "LOO",
  fixed.block = NULL,
  design.block = NULL,
  canonical = TRUE,
  ...
)

Arguments

formula
Model formula accepting a single response (block) and predictor block names separated by + signs.

ncomp
The maximum number of ROSA components.

Y.add
Optional response(s) available in the data set.

common.comp
Automatically create all combinations of common components up to length common.comp (default = 1).

data
The data set to analyse.

subset
Expression for subsetting the data before modelling.

na.action
How to handle NAs (no action implemented).

scale
Optionally scale predictor variables by their individual standard deviations.

weights
Optional object weights.

validation
Optional cross-validation strategy "CV" or "LOO".

internal.validation
Optional cross-validation for block selection process, "LOO", "CV3", "CV5", "CV10" (CV-number of segments), or vector of integers.

fixed.block
integer vector with block numbers for each component (0 = not fixed) or list of length <= ncomp (element length 0 = not fixed).

design.block
integer vector containing block numbers of design blocks

canonical
logical indicating if canonical correlation should be use when calculating loading weights (default), enabling B/W maximization, common components, etc. Alternatively (FALSE) a PLS2 strategy, e.g. for spectra response, is used.

...
Additional arguments for cvseg or rosa.fit

Value

An object of classes rosa and mvr having several associated printing and plotting methods.

References

Liland, K.H., Næs, T., and Indahl, U.G. (2016). ROSA - a fast extension of partial least squares regression for multiblock data analysis. Journal of Chemometrics, 30, 651–662, doi:10.1002/cem.2824.

See also

Examples

data(potato)
mod <- rosa(Sensory[,1] ~ ., data = potato, ncomp = 10, validation = "CV", segments = 5)
summary(mod)

#> Data: 	X dimension: 26 3946 
#> 	Y dimension: 26 1
#> Fit method:
#> Number of components considered: 10
#> 
#> VALIDATION: RMSEP
#> Cross-validated using 5 random segments.
#>        (Intercept)  1 comps  2 comps  3 comps  4 comps  5 comps  6 comps
#> CV           1.778    1.068   1.0064    1.013   0.8680    1.075    1.208
#> adjCV        1.778    1.030   0.9657    0.968   0.8353    1.020    1.140
#>        7 comps  8 comps  9 comps  10 comps
#> CV       1.260    1.237    1.513     1.642
#> adjCV    1.192    1.185    1.425     1.537
#> 
#> TRAINING: % variance explained
#>               1 comps  2 comps  3 comps  4 comps  5 comps  6 comps  7 comps
#> X               27.82    72.05    82.50    83.10    84.88    87.09    87.24
#> Sensory[, 1]    76.53    83.36    84.21    87.91    88.06    88.18    88.17
#>               8 comps  9 comps  10 comps
#> X               88.46    91.18     92.81
#> Sensory[, 1]    88.47    88.78     88.78